jobs/sentiment_call/requirements.txt
google-cloud-bigquery
google-cloud-pubsub
google-cloud-language
